Allow for any range of values to be plotted
Previously, I was using a confusing system which offered little in the
way of changing the range of values plotted etc. Now, I can plot any
range of real and imaginary values with ease, and hopefully the code is
now clearer in how it works. It now computes the increase in value for
each axis per pixel moved. Then it computes whether that point is the in
Set, and plots the colour accordingly.