Started using Tkinter for displaying plot
I was previously using PIL, but this did not give me very much control
over the actual window. I eventually want to build something similar to
Matplotlib, which is capable of zooming in/out of a selected region, and
of panning the plot. For this purpose, I switched to Tkinter.