Created 'Diagram' class
This class is intended represent the Argand Diagram which is used to
display the plot. Currently, this class has attributes to store the
dimensions of the window for which it has been formatted, the ranges of
the real and imaginary axes used in the diagram, and the diagram's
deltas in the x and y directions. It also includes a method for
calculating new deltas if the ranges of the axes have been updated. I
have included this to make it easier later to allow for zooming etc. (to
be honest, this is my main reason for creating this class).