Mouse pos now only updates label if over diagram
Before, the label to display the mouse's current position was updated
even if the mouse was moving over the label rather than over the
diagram. This was clearly not what I wanted, so I therefore have bound
the motion event to just the diagram widget, rather than the entire
window.