Make append_to_zoom_list take values to append
Before, it was operating in a dodgy way where it was implicit that the
values for the ranges had already been updated, and the role of this
function was merely to update zoom_list with the new values. However, as
we have changed how we define the ranges, this approach no longer works.
Therefore we now pass the values to be appended to this function, and
let it append them.