Fix y-axis flipping
I was calculating the points from the bottom left corner to the top
right corner, but drawing them from the top left corner to the bottom
right corner. This led to the display being flipped in the y-axis. Now,
I am calculating and drawing from the bottom left corner to the top
right corner, fixing the issue.